pyproject.toml
[project]
name = "wmca"
version = "0.1.0"
requires-python = ">= 3.11"
dependencies = [
    "torch",
    "numpy",
    "wandb",
    "python-dotenv",
]

[project.optional-dependencies]
cuda = ["triton"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
